requirements.txt
Django==2.2.13
django-crispy-forms==1.7.2
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
